requirements.txt
flask
gunicorn
pytesseract
pillow
pdf2image